added fetching IMDB ratings

+++ src/main/java/com/cgi/lauri/movieRecommender/dto/MovieDto.java
@@ -25,5 +25,5 @@ public class MovieDto {
     private Integer minimumAge;
     //private List<Screen> screens;
     private Set<MovieRating> ratings;
-    private Integer imdbRating;
+    private Double imdbRating;
 }
